requirements.txt
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
transformers>=4.41.0,<5.0.0
scikit-learn>=1.4.0
streamlit>=1.36.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
pytest>=8.0.0
